+++ pomace/config.py
@@ -37,6 +37,7 @@ class Site:
 
 @datafile("./pomace.yml", defaults=True)
 class Settings:
+    framework: str = "splinter"
     browser: Browser = field(default_factory=Browser)
     url: str = ""
     aliases: Dict[str, str] = field(default_factory=dict)
@@ -51,7 +52,7 @@ class Settings:
         return self.get_secret(name) or ""
 
     def get_secret(self, name, *, _log=True) -> Optional[str]:
-        domain = URL(shared.browser.url).domain
+        domain = URL(shared.client.url).domain
         for site in self.secrets:
             if site.domain == domain:
                 for secret in site.data:
@@ -62,7 +63,7 @@ class Settings:
         return None
 
     def set_secret(self, name, value):
-        domain = URL(shared.browser.url).domain
+        domain = URL(shared.client.url).domain
         self._ensure_site(domain)
         site: Site = self._get_site(domain)  # type: ignore
         for secret in site.data:
@@ -73,7 +74,7 @@ class Settings:
             site.data.append(Secret(name, value))
 
     def update_secret(self, name, value):
-        domain = URL(shared.browser.url).domain
+        domain = URL(shared.client.url).domain
         site = self._get_site(domain)
         if site:
             for secret in site.data: